split large modules & annotation in search
