+++ genomepy/plugins/star.py
@@ -2,16 +2,9 @@ import os
 
 from loguru import logger
 
+from genomepy.files import bgzip_and_name, gunzip_and_name, gzip_and_name
 from genomepy.plugin import Plugin
-from genomepy.utils import (
-    mkdir_p,
-    rm_rf,
-    cmd_ok,
-    run_index_cmd,
-    gunzip_and_name,
-    bgzip_and_name,
-    gzip_and_name,
-)
+from genomepy.utils import mkdir_p, rm_rf, cmd_ok, run_index_cmd
 
 
 class StarPlugin(Plugin):
