+++ tests/test_07_provider_ensembl.py
@@ -2,22 +2,25 @@ import requests
 
 import pytest
 
+import genomepy.providers.ensembl
 from tests import linux, travis
 
 
-def test_ensemblprovider__init__(ensembl):
+def test_ensemblprovider(ensembl):
     assert ensembl.name == "Ensembl"
     assert ensembl.taxid_fields == ["taxonomy_id"]
 
 
-def test__request_json(ensembl):
-    divisions = ensembl._request_json("https://rest.ensembl.org/", "info/divisions?")
+def test_request_json(ensembl):
+    divisions = genomepy.providers.ensembl.request_json(
+        "https://rest.ensembl.org/", "info/divisions?"
+    )
     assert isinstance(divisions, list)
     assert "EnsemblVertebrates" in divisions
 
     # test not r.ok
     with pytest.raises(requests.exceptions.HTTPError):
-        ensembl._request_json("https://rest.ensembl.org/", "error")
+        genomepy.providers.ensembl.request_json("https://rest.ensembl.org/", "error")
 
 
 def test__get_genomes(ensembl):
@@ -35,17 +38,17 @@ def test__get_genomes(ensembl):
 def test_genome_info_tuple(ensembl):
     t = ensembl._genome_info_tuple("KH")
     assert isinstance(t, tuple)
-    assert t[2:4] == ("Ciona intestinalis", "7719")
+    assert t[0:4] == ("KH", "GCA_000224145.1", 7719, True)
 
 
 def test_get_version(ensembl):
-    v = ensembl.get_version("https://rest.ensembl.org/", True)
+    v = ensembl.get_version("https://rest.ensembl.org/", vertebrates=True)
     assert v.isnumeric()
     assert int(v) > 100
 
     v = ensembl.get_version("https://rest.ensembl.org/")
     assert v.isnumeric()
-    assert int(v) > 48
+    assert int(v) > 50
 
 
 @pytest.mark.skipif(travis and linux, reason="FTP does not work on Travis-Linux")
@@ -89,28 +92,28 @@ def test_get_genome_download_link(ensembl):
 
 
 @pytest.mark.skipif(travis and linux, reason="FTP does not work on Travis-Linux")
-def test_get_annotation_download_link(ensembl):
+def test_get_annotation_download_links(ensembl):
     # non vertebrate
-    link = ensembl.get_annotation_download_link("TAIR10", **{"version": 46})
+    links = ensembl.get_annotation_download_links("TAIR10", **{"version": 46})
     expected_link = (
         "ftp://ftp.ensemblgenomes.org/pub/plants/release-46/"
         "gtf/arabidopsis_thaliana/Arabidopsis_thaliana.TAIR10.46.gtf.gz"
     )
-    assert link == expected_link
+    assert links[0] == expected_link
 
     # vertebrate
-    link = ensembl.get_annotation_download_link("GRCz11", **{"version": 98})
+    links = ensembl.get_annotation_download_links("GRCz11", **{"version": 98})
     expected_link = (
         "ftp://ftp.ensembl.org/pub/release-98/gtf/"
         "danio_rerio/Danio_rerio.GRCz11.98.gtf.gz"
     )
-    assert link == expected_link
+    assert links[0] == expected_link
 
     # vertebrate: latest version
     version = ensembl.get_version("https://rest.ensembl.org/", True)
-    link = ensembl.get_annotation_download_link("GRCz11", **{"version": version})
+    links = ensembl.get_annotation_download_links("GRCz11", **{"version": version})
     expected_link = (
         f"ftp://ftp.ensembl.org/pub/release-{version}/"
         f"gtf/danio_rerio/Danio_rerio.GRCz11.{version}.gtf.gz"
     )
-    assert link == expected_link
+    assert links[0] == expected_link
