+++ tests/test_08_provider_ucsc.py
@@ -2,8 +2,10 @@ import os
 from shutil import copyfile
 from tempfile import TemporaryDirectory
 
+import pytest
 
-def test_ucscprovider__init__(ucsc):
+
+def test_ucscprovider(ucsc):
     assert ucsc.name == "UCSC"
     assert ucsc.taxid_fields == ["taxId"]
 
@@ -20,12 +22,11 @@ def test__get_genomes(ucsc):
 
 def test__search_accession(ucsc):
     assert list(ucsc._search_accession("not_an_id")) == []
-    assert list(ucsc._search_accession("GCA_000004335.1")) == ["ailMel1"]
+    assert next(ucsc._search_accession("GCA_000004335.1")) == "ailMel1"
 
 
 def test_assembly_accession(ucsc):
-    genome = ucsc.genomes["sacCer3"]
-    accession = ucsc.assembly_accession(genome)
+    accession = ucsc.assembly_accession("sacCer3")
 
     assert accession.startswith("GCA_000146045")
 
@@ -33,7 +34,7 @@ def test_assembly_accession(ucsc):
 def test_genome_info_tuple(ucsc):
     t = ucsc._genome_info_tuple("sacCer3")
     assert isinstance(t, tuple)
-    assert t[2:4] == ("Saccharomyces cerevisiae", "559292")
+    assert t[0:4] == ("sacCer3", "GCA_000146045.2", 559292, [False, True, True, False])
 
 
 def test_get_genome_download_link(ucsc):
@@ -78,33 +79,48 @@ def test__post_process_download(ucsc):
                 assert "a" not in line
 
 
-def test_get_annotation_download_link(ucsc):
+def test_get_annotation_download_links(ucsc):
     # any GTF format annotation
     genome = "sacCer3"
-    link = ucsc.get_annotation_download_link(genome)
-    assert genome in link
-    assert link.endswith(".gtf.gz")
+    links = ucsc.get_annotation_download_links(genome)
+    assert genome in links[0]
+    assert links[0].endswith(".gtf.gz")
 
+    # any TXT format annotation (no GTF available)
+    genome = "xenTro2"
+    links = ucsc.get_annotation_download_links(genome)
+    assert genome in links[0]
+    assert links[0].endswith(".txt.gz")
+
+    # no annotation available
+    assert ucsc.get_annotation_download_links("apiMel1") is None
+
+
+def test_get_annotation_download_link(ucsc):
     # specific GTF annotation type
+    genome = "sacCer3"
     link = ucsc.get_annotation_download_link(
         genome, **{"ucsc_annotation_type": "NCBI_refseq"}
     )
     assert genome in link
     assert link.endswith("ncbiRefSeq.gtf.gz")
 
-    # any TXT format annotation (no GTF available)
-    genome = "xenTro2"
-    link = ucsc.get_annotation_download_link(genome)
-    assert genome in link
-    assert link.endswith(".txt.gz")
-
     # specific TXT annotation type (no GTF available)
+    genome = "xenTro2"
     link = ucsc.get_annotation_download_link(
         genome, **{"ucsc_annotation_type": "UCSC_refseq"}
     )
     assert genome in link
     assert link.endswith("refGene.txt.gz")
 
-    # non-existing annotation type
-    link = ucsc.get_annotation_download_link(genome, **{"ucsc_annotation_type": "UCSC"})
-    assert link is None
+    # annotation type non-existing for this genome
+    with pytest.raises(FileNotFoundError):
+        ucsc.get_annotation_download_link(genome, **{"ucsc_annotation_type": "UCSC"})
+
+    # annotation type non-existing
+    with pytest.raises(ValueError):
+        ucsc.get_annotation_download_link(genome, **{"ucsc_annotation_type": "what?"})
+
+    # no annotation available
+    with pytest.raises(FileNotFoundError):
+        ucsc.get_annotation_download_link("apiMel1")
